[class] refactor : 클래스 상세 페이지 리팩토링
